third_party/blink/web_tests/platform/generic/external/wpt/html/cross-origin-opener-policy/tentative/restrict-properties/iframe-popup.https_5-6-expected.txt

This is a testharness.js-based test.
FAIL COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with same site iframe, same site popup with window_open assert_equals: Iframe has dom access to the popup? expected "false" but got "true"
PASS COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with same site iframe, same site popup with anchor
PASS COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with same site iframe, same site popup with form
FAIL COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with cross origin iframe, same site popup with window_open assert_equals: Iframe has cross origin access to the popup? expected "false" but got "true"
PASS COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with cross origin iframe, same site popup with anchor
PASS COOP: restrict-properties to popup COOP: restrict-properties via an iframe, with cross origin iframe, same site popup with form
Harness: the test ran to completion.
